Numerical vs categorical X-axis
Adds logic to adjust the position of x-axis labels (above or below) the
plot based on whether the data is categorical (above) or numerical
(below). Does this by checking the type of the first value of the
X-axis values.

Adds the logic set up as a stub in commit c68787e.

Additional test case is added to confirm this new functionality